backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Caching & Queue
redis==5.0.1
celery==5.3.6

# HTTP Client
httpx==0.26.0
aiohttp==3.9.1

# Environment
python-dotenv==1.0.0

# Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Utilities
python-dateutil==2.8.2
orjson==3.9.10
structlog==23.2.0

# Stripe for billing
stripe==7.8.0

# Testing (dev only)
pytest==7.4.4
pytest-asyncio==0.21.1
